moving LanguageObject to another file
